merge

+++ kernel/include/linux/syscalls.h
@@ -861,6 +861,6 @@ asmlinkage long sys_finit_module(int fd, const char __user *uargs, int flags);
 asmlinkage long sys_seccomp(unsigned int op, unsigned int flags,
 			    const char __user *uargs);
 
-asmlinkage int sys_ptree(struct prinfo __user *buf, int __user *nr);
+asmlinkage long sys_ptree(struct prinfo __user *buf, int __user *nr);
 
 #endif
